model/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
ultralytics
librosa
opencv-python
numpy
Pillow
pyngrok
requests
pydantic
